Granular media permissions
This commit adds `READ_MEDIA_AUDIO` permission to the manifest and keeps `READ_MEDIA_VIDEO` permission, which are required to access audio and video files for Android 13 and higher. The `READ_EXTERNAL_STORAGE` permission is still included with a max SDK version of 32 to handle devices running older android versions.