Expose player state and player instance
This commit exposes the `player` instance in `PlayerState` class to be accessible from outside. It also expose the `isPlaying` state from `PlayMediaUseCase` and access it in `MainViewModel` using `isPlaying()` function. In the `MainScreen` the `isPlaying` is now provided from `MainViewModel` instead of a local `var` state, this ensure a proper and centralized `isPlaying` value that reflects the current state.